requirements.txt
Package           Version
----------------- -------
aiohappyeyeballs  2.6.1
aiohttp           3.12.15
aiosignal         1.4.0
async-timeout     5.0.1
attrs             25.3.0
discord.py        2.6.3
dotenv            0.9.9
frozenlist        1.7.0
idna              3.10
multidict         6.6.4
pip               22.0.2
propcache         0.3.2
psycopg           3.2.9
python-dotenv     1.1.1
setuptools        59.6.0
typing_extensions 4.15.0
yarl              1.20.1
